analise_releases/assistant/requirements.txt
chainlit>=1.1.0
groq>=0.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
chromadb>=0.4.0
pypdf>=4.0.0
sentence-transformers>=3.0.0
